test: add /api/tasks route and useTasks hook coverage
- Add list_tasks_app() helper in tests/common.rs routing both /tasks
  and /api/tasks to list_tasks handler
- Add api_tasks_alias_returns_ok test verifying GET /api/tasks → 200
- Add tasks_and_api_tasks_return_same_data parity test
- auth.rs already covers /api/tasks via build_router real-router tests
- queries.test.tsx already covers useTasks hook → /api/tasks fetch

Fixes reviewer concern: a typo or removal of the /api/tasks route now
fails CI rather than silently recreating the dashboard 404.
