+++ crates/harness-server/src/http/tests/mod.rs
@@ -1,0 +1,4 @@
+mod api;
+mod auth;
+mod common;
+mod webhook;
